Create LAST_CHUNK message
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
Create `LAST_CHUNK` message to tell client stop reciving the data chunks and deserialize the data.
The added `isLast` and `seializedLastChunk` function in this diff is not used in production.
In the next diff, we add a new capability to version seven to send the data in chunks, and use these two functions.

Reviewed By: jdelliot

Differential Revision: D73400597

fbshipit-source-id: e8237aaf2753915894ad57f60609aec37b2e9101