minor refactoring in TakeoverClient
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
Made some minor refactoring in `TakeoverClient` to prepare it for the next diff changes. No logic change in this diff

Reviewed By: jdelliot

Differential Revision: D73221179

fbshipit-source-id: ad43ca62cd723a6eba7cccb6129837fddeaedd18