record: improve error handling when restoring files
Summary:
If we get an error when restoring files back to the working copy (typically after some "--interactive" command), previously we would bail out and silently stop restoring files. Now we will continue and try restoring the remaining files.

I also added an error message that can help the user manually recover the file content should a file fail restoring.

In particular, a user hit this when a file couldn't be restored because a process had the working copy file open (Windows).

Reviewed By: quark-zju

Differential Revision: D73464845

fbshipit-source-id: 6c4ad4633c37fb0391711edb4161a14f6cf781db