Folded non-streaming and streaming EdenFS clients into a single struct/impl
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Having these two separate structs/impls was resutling in some challenging issues with testing and cases where a single command or API used both. Lets fix that be merging them. In a subsequent diff, we can take a stab a sharing code between the different `with_XXX` methods, but only if it cleans up the code in a useful manner.

Reviewed By: genevievehelsel

Differential Revision: D73380905

fbshipit-source-id: 6d40bdd034208c46b9bf629fc51d1dd319540ac3