checkout: add flag to propagate checkout errors
Summary:
Add a new experimental.propagate-checkout-errors flag which causes eden's "checkout" operation to propagate errors as top level checkout errors instead of reporting them as per-path "conflict" errors.

We only propagate errors from the "core" checkout operation. We can also get per-path errors during "secondary" operations such as invalidation, but we don't those errors yet (other than as "conflicts"). Invalidation failures should probably also propagate, but let's start with the really gimme errors (e.g. "couldn't check out anything because all tree fetches are failing").

The problem with the existing behavior is that, even though the error is propagated to Sapling, Eden pretends that the checkout operation succeeded, which can leave you in an inconsistent state with incorrect file contents. Even if Sapling wanted to fix the mount, it can't. Eden doesn't think anything is wrong, and pending user modifications could be mixed in with didnt-actually-check-out-so-are-modified modifications. Resolve by making Eden enter the interrupted checkout state on error. This causes "status" to report and error, telling the user how to complete the checkout. Although annoying, this at least avoids potential data lass and incorrect commit content.

In the future, I will make Sapling automatically complete interrupted checkouts.

The context here is we saw a spurt of network errors during checkout operations as we rolled out new load shedding on the server. This brought the bad checkout behvaior to the forefront for multiple users.

Note we have an existing experimental.abort-on-eden-conflict-error config in Sapling. This was added to at least bail out on errors in CI (even though the repo can't necessarily be recovered). I determined the problem is not possible to fix within Sapling, so instead I moved the bail-on-error logic into Eden. Once this new behavior is rolled out we can get rid of experimental.abort-on-eden-conflict-error.

Reviewed By: genevievehelsel

Differential Revision: D73401957

fbshipit-source-id: 1eebcbdc491d705abd00f340aa740c72a1ca327e