support Rust failpoints for tests
Summary:
The "fail" crate lets you force instrumented codepaths to return errors via the FAILPOINTS env var. Sapling uses this, but previously it wasn't possible to use via eden. Make it work by propagating the FAILPOINTS env var to eden, and triggering the failpoint setup in the backingstore.

In the future we may want to support setting failpoints over thrift so we don't need to restart eden.

Reviewed By: quark-zju

Differential Revision: D73401960

fbshipit-source-id: c67c59b3a95698e7f3acfb4ca1cbcce3eb3d3bbf