testutil: add global failpoints initializer
Summary:
In certain cases (such as within long running EdenFS process) the FailScenario (which contains MutexGuard) is not convenient to set up and hold since it isn't Send. Add a simpler `setup_global_fail_points` which activates the failpoints based on $FAILPOINTS without needing to retain any values.

It is still possible to clean up the fail points if necessary, but I don't need that now, so we can add later if needed.

Reviewed By: quark-zju

Differential Revision: D73401955

fbshipit-source-id: 6b89063358effbd8b0b9045582907c2a503e10d6