Remove the remaining streaming connection specifc code
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

We are now able to unify the rest of the code to have one Thrift service client. This means we can get rid of things like `with_streaming_thrift`.

Reviewed By: MichaelCuevas

Differential Revision: D73468548

fbshipit-source-id: 54bff67326226cc23b30673447c35869036e7b79