Enhanced get_changes_since unit test
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Added a support for creating result structs ergonomicall and comparing them for testing. Added a tiny bit more data to the test. More to follow.

Reviewed By: MichaelCuevas

Differential Revision: D73479873

fbshipit-source-id: e7ec4b6f2c8c4e9cd8cee58c8ea79a1b702310d5