Begin using StreamingEdenService in place of EdenService
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

StreamingEdenService is a supertrait of EdenService and others. Knowing that, we can simplify the code to only use StreamingEdenService. This diff begins that process by replacing all usage of EdenService with StreamingEdenSerivce. We also had to include mocking StreamingEdenService to get test code to compile.

# Next Steps:

* We will also add support for mocking streaming client and validate that can be done in the same manner.

Reviewed By: MichaelCuevas

Differential Revision: D73457874

fbshipit-source-id: e5e6a507a5c6113236c383c3886aa3b961301b12