Add some comments regarding mock maintenance for EdenFS Thrift APIs
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

One side-effect of adding test support for EdenFS Thrift APIs from the client is that we will now need to make sure we update the mocks when introducing a new method or a breaking API change. This is not necessarily a bad thing, but something we will want to make clear. Added comments to both of the thrift defintion files and the mock file.

Differential Revision: D73521645

fbshipit-source-id: 4328de2bee06ba8effa69ffb7985ff0e2c9cb90c