send empty bundle-list upon failure
Summary: When we fail to generate a bundle-list, the whole git command fails, but when we send an empty bundle list, nothing happens - git just sees it's empty and does nothing. In case of a sever-side failure, let's just fail silently and send an empty bundle-list to the client. When we fail silently we have to log it, otherwise there's no way of knowing we're failing, so I'm also adding logging here.

Reviewed By: RajivTS

Differential Revision: D73188354

fbshipit-source-id: ebc80c2a2b423e640fe175d4c8b067b9cc7db9d7