Enable loading of selective set of repos
Summary:
Since SCS runs for all repos at Meta (Hg and Git), its been configured to load only the shallow sharded repos at start up. In practice this means loading all the Hg repos and letting ShardManager provide the rest of the Git repos during runtime in production.

However, this model doesn't quite work for local testing if someone wants to test SCS changes for a particular Git repo. This diff adds a flag (and associated warning) that allows users to load a specific set of repos.

To do this, users have to use the `--load-all-repos-in-tier` flag along with `--filter-repos` argument specifying the regex identifying the set of repos they want to load at startup

Reviewed By: singhsrb

Differential Revision: D73778120

fbshipit-source-id: 57f1ac0f9434d2b99a6b540f847cadaead46b100