walkdetector: support advancing BFS depth boundary
Summary:
Introduce a new heuristic to expand the depth of an in-progress walk. Basically, increase depth to depth+1 when we observe N children of the walk root with descendants at depth+1. The intuition is we don't want to advance the depth if only one subtree is being walked (instead, we should track that one subtree as a separate walk).

For example:
```
     A
   /   \
  B      C
 / \    / \
D   E  F   G
```

If we have walk(root=A, depth=1) (which includes B and C) and see D being accessed, we don't want to assume E, F, and G are needed yet (due to exponential growth of next depth). Similarly, if D and E are accessed, we might not want to increase the depth since F and G might not be needed. However, if we see D and F get loaded, that is evidence that both B and C are walked at the next depth, so we increase depth to 2.

Reviewed By: zzl0

Differential Revision: D73407855

fbshipit-source-id: 6669134ac77ccff2de11667b4f2266c7d5a5b23f