walkdetector: merge "dirs" into the WalkNode tree
Summary: I've realized it doesn't make sense to have two separate data structures for directory metadata and walk metadata. This is simpler and more useful. In particular, previously we would clear out the directory hints as soon as we converted a `Dir` into a `Walk`. Now, the hints will naturally live on alongside the `Walk` until the containing `WalkNode` itself is dropped (when consumed by a higher up walk).

Reviewed By: zzl0

Differential Revision: D73407869

fbshipit-source-id: 9d2afa4f492e8ef10a323139e8e4d80c2f2b617a