walkdetector: add directory size hints
Summary:
Add a `read_dir` method which witnesses directory reads. It remembers the number of children files/dirs to use later when deciding if things are being walked or not.

In practice this information will be best effort since EdenFS (as the filesystem) won't see directory reads if they are served from the kernel cache. If this information turns out to be very important (and not present often enough), we can consider making the walk detector "actively" query for directory metadata, but I'd like to avoid that if possible since it makes things more complicated.

Reviewed By: zzl0

Differential Revision: D73407877

fbshipit-source-id: 524b2d236eabaf91c11df6272eb295fbf8e348f8