Check JK for WBC update mode in every iteration
Summary: I initially kept this JK out of the loop thinking we would generate considerable strain on the JK backend by querying it in every iteration of the loop. I have since learned that JKs are quite good at managing this kind of query volume and we would definitely like the luxury of reverting this change at a moment's notice like in S514675. This diff does that.

Reviewed By: lmvasquezg

Differential Revision: D73784788

fbshipit-source-id: 6af5db635516316c45bd057da91ca4211e7c2c7f