Replace isChunked with isFirstChunk
Summary:
## Context:
EdenFS crash during graceful restart when the `TakeoveData` is large. One solution for this crash is to spilit the `TakeoverData` into 1 GB chunks, and send data over socket in chunks. More explanations and pseudocode is in this [post](https://fb.workplace.com/groups/1708850869939124/permalink/1952362862254589/)

## This diff:
`isChunked` dosen't check the content of the message, it only checks the capability of the message in the header and detects if the message is sent in chunks.
On the other hand, if the server faced to any error in sending the data, it uses `sendError` to wrap the error message in a socket message and send it over the socket.
https://www.internalfb.com/code/fbsource/[1f88cfd5b007]/fbcode/eden/fs/takeover/TakeoverServer.cpp?lines=188
Therefore, we could have a message in the client side which has the correct header with the error.
This could cause issues for clients that need to read data in chunks. Because the error message has the correct header and correct capability, and `isChunked` return true for error messages.
To fix this issue, this diff replaces `isChunked` with `isFirstChunk`. In the next diff, the server use `serliazeFirstChunk` to send a small message for the fist chunk. This first chunk message should send before the first chunk.

Reviewed By: jdelliot

Differential Revision: D73514373

fbshipit-source-id: be56578bb9b38342a2a0810f2c566a0ac0c648b9