Include Tree aux data within the response from sapling native backing store
Summary:
Modify ffi so sapling::Tree so it contains aux data besides entries.

We used to discard the aux data while fetching tree from Sapling. With the local store ready for Tree aux data, we should take advantage and store the aux data retrieved from Sapling when a tree is queried.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72425830

fbshipit-source-id: 3c9419427192900da001d2bfa86857a0ab2bb9d9