Enable finer JK control for bonsai tag mapping caching
Summary: In S514675, we realized that turning off scribe-based update of our caches required not only JK change but also a restart of the servers. This isn't ideal because its disruptive to the service and more importantly it adds a considerable amount of delay in addressing the problem. With this diff, we enable immediate rollback of similar changes for bonsai tag mapping caching just through JK change, although at this level it cannot be configured per-repo but instead applies to the feature as a whole.

Reviewed By: YousefSalama

Differential Revision: D73862867

fbshipit-source-id: f41cd70501fdecf15ff44f2e679871329583924b