Fix a crash in folly::dynamic
Summary:
I found some edge cases that can end in a crash in auto unmount recovery:
1- If we have two accidental unmounted repos, the recovery crashes and cannot remount two repos
2- If the mount path of an unmounted repo is removed from the disk, the recovery crashes and cannot remount the repo.

The root cause of both crashes is in the way that `client` which is a `folly::dynamic` is passed into the lambda function. The `client.first` and  `client.second` are string out of the lambda function scop. However, inside the lambda these are not string. To fix this issue, this diff copy the repo name into a new string `repoName` and pass this string into the lambda function instead of the `folly::dynamics`

The manual tests show that the crash is fixed with the change.

#thanks genevievehelsel for debugging this crash with me.

Reviewed By: genevievehelsel, jdelliot

Differential Revision: D73817423

fbshipit-source-id: 8e8cf66d1e041d72c4c4a08264365d0e5039a031