Move some tests to treehugger oncall
Summary:
I'm oncall and still getting assigned tasks for test issues, which should go to the treehugger, e.g. T191383599.

There are some tests in that task that are already in the treehugger oncall, so IDK why they are in a task assigned to me, but while looking I found a few that still need to be updated.

I just ran find and replace in `eden/mononoke/tests` directory

Reviewed By: RajivTS

Differential Revision: D73915377

fbshipit-source-id: 0461d6917af47edb27a0b959c41d58bf8f943cd7