Log bookmark movements for backsynced repo
Summary:
For push-redirected repos, we log bookmark movement only for the source (large) repo (e.g. `fbsource` in case of `fbsource -> whatsapp/*` sync) while ignoring the corresponding bookmark move in the target (small) repo. This was fine until we decided to update our WBC refresh logic to rely on bookmark movement logging to scribe which resulted in S514675

This diff enables logging of bookmark movement for push-redirected repos so scribe-tailing mode of WBC update can be enabled for such repos (e.g. whatsapp family of repos)

Reviewed By: gustavoavena

Differential Revision: D73861233

fbshipit-source-id: 8f06846bf5693b85b1dc7b92bb2ac019fe74fde0