enabling passing aux data with cached tree
Summary:
more code refactoring

This code modification serves the following purpose: it localizes the issue for fast
and slow path, now the conversion to a sapling blob and converting back into
the ManifestTreeEntry happens in a single place.

Also this diff should make aux data propagation work on the fast path.

Reviewed By: muirdm

Differential Revision: D73929551

fbshipit-source-id: 1d23fe0c3f603c30514820cfa4af6b02249b71ee