commitcloud: add a config to ignore certain local bookmarks
Summary:
Use case: for a "warm up" commit that is only "warmed up" on a single host, not
globally. It's nice to use a local bookmark to indicate that the commit is
warmed up. The bookmark name can be used in revset (namespaces), and shows up
in ISL. But the bookmark shouldn't be synced to other hosts.

Reviewed By: liubov-dmitrieva

Differential Revision: D73946098

fbshipit-source-id: d2048dde0c14321bf7fb6dba7212267809caff5e