print daemon's environment variables
Summary:
# This diff

Adds a section to `eden rage` output that prints out the environment variables that are present for the daemon process. This differs from the env vars that are present on the host, so the output is updated to reflect that

# Context

When investigating S513355, we could have narrowed down the root cause much earlier if we knew that the env vars printed in the rage were NOT the env vars that were in effect for the daemon.

To avoid that confusion in the future, I've updated the env var output to reflect that the existing list contains the env vars for the host. I've also added a section to print the env vars for the daemon, which differs substantially from the env vars for the host.

Reviewed By: jdelliot, genevievehelsel

Differential Revision: D74032644

fbshipit-source-id: e3090b8e71f6ab80e2a85f734c84ba4d35f2628a