F401 in windows_fsck_test.py
Summary:
## Instructions about RACER Diffs
This diff is removing the unused imports from Code Quality Score from [Monetization codehub](https://fburl.com/quality/58ozwacb):
 - Accept and ship the diff (racer do not do auto land)
 - [**Recommended**]Commandeer and land this diff.

If you are happy with the changes, please accept and ship, or commandeer it if minor edits are needed.
If you are not happy with the changes, please comment on the diff with clear actions and send it back to the author. Racer will pick it up and re-generage.
If you really feel the Racer is not helping with this change (alas, some complex changes are hard for AI) feel free to abandon this diff.

This diff was pre-created by Racer AI agent for your convenience on top of T223415841. More details about the program see this [post](https://fb.workplace.com/groups/1177219810659335/permalink/1198503531864296/). For questions or suggestions please post in [Racer Users and Feedbacks](https://fb.workplace.com/groups/1112800313955334) group.
 ---
## Summary of Change
Perfect! I've successfully addressed the F401 issue by:

- Removing the unused import `unittest` (identified in the task)
- Also removing the unused import `sys` that was found during linting
- Running linting to confirm the changes fixed the issues
- Creating a local commit with the changes

Here is a summary of the changes I made:

- [RACER] Remove unused imports in windows_fsck_test.py
- Removed the unused imports: unittest and sys
- Running arc lint confirmed the fixes

Reviewed By: genevievehelsel

Differential Revision: D74232654

fbshipit-source-id: b2117d4f213098791ed80d1b600a11d28ead2e11