adding option to disable scrolling when opening inset view
Summary:
**Context**

I notice that sometimes when the inset is on the edge of the screen + when line calculation is wonky and not 100%, opening an inset would scroll the user.

I will look at the line adjustment logic again but I think when toggling the inset open in editor, we shouldnt scroll the user.

Reviewed By: evangrayk

Differential Revision: D74273975

fbshipit-source-id: 466fb3e3254514cd60efb16086ecad2b7b353b54