Register waker when returning Poll::Pending from Mononoke Git read path
Summary:
When we first hand crafted the stream for fetching trees and blobs in Mononoke Git server, we realized that an unoptimized build could make it so that we reach the end of the stream body without even a single future being ready. To address this case, we put in a `Poll::Pending` to ensure that the stream does not incorrectly assume itself to get completed but instead marks itself for bring re-polled.

A mistake that we made back then was to not register a waker for the pending stream. So far, whenever this situation was hit there was an external entity manually re-polling the stream ensuring that we make progress. However, recently we encountered a peculiar situation where this stopped being true.

Cloning `aosp/platform/external/apache-xml` repo from Mononoke would get stuck at 73% and remain stuck there forever. This repo has a particular commit history where there are hundreds of commits with the same root tree, i.e. they introduce no changes. This triggers a situation where we get to the end of the stream body without all changesets being processed and hence return `Poll::Pending`. However in this case, there is no external entity re-polling the stream which leads to the stream and hence the clone getting stuck forever.

The solution here was to register the waker to immediately wake this stream up cause a second pass would result in more values.

Differential Revision: D74397719

fbshipit-source-id: af2ba16a4369eb6d0a055948ed76191358bb022a