Add a varint for LazyTree which contains the aux data
Summary:
## Context
Currently, when a tree is fetched from hgcache, we don't include the tree aux data with it even when we have the aux data cached locally already. It might be useful for some use cases where the tree aux data is included in the tree response so the user does not need another round trip asking for it.

## This Diff
We introduce a new variant `IndexedLogWithAuxData` for `LazyTree` enum where we can put the (optional) tree aux data.
For `LazyTree`, the only function we care to implement for `IndexedLogWithAuxData` is `aux_data()` since the rest is not going to be used for our use case.

## Next Step
Modify `TreeStore::get_local_tree_direct` so it returns an instance of `ScmStoreTreeEntry` instead of `manifest_tree::store::Entry`. We will be using the new variant to construct `LazyTree` for `ScmStoreTreeEntry`.

Reviewed By: muirdm

Differential Revision: D74037843

fbshipit-source-id: 6da0cecc153f6a38422f877a3f330c87dbb33509