Wrap raw tree entry within ScmStoreTreeEntry
Summary:
## Context
Currently, when a tree is fetched from hgcache, we don't include the tree aux data with it even when we have the aux data cached locally already. It might be useful for some use cases where the tree aux data is included in the tree response so the user does not need another round trip asking for it.

## This Diff
Modify TreeStore::get_local_tree_direct so it returns an instance of ScmStoreTreeEntry instead of manifest_tree::store::Entry. We use the new variant to construct LazyTree for ScmStoreTreeEntry.

## Next Step
Modify the structure of EdenFS Tree so it includes the aux data.

Reviewed By: liubov-dmitrieva

Differential Revision: D74050122

fbshipit-source-id: bf54a338ccfab0bf99e2e7d8e9b9ad3b7b96341e