pipe object paths to SaplingNativeBackingStore
Summary:
I want to use tree/blob paths to detect when eden is being walked. The path seems to be available in the ProxyHash, but that is not propagated into Sapling land at all. Let's start by adding it to the SaplingRequest which gets passed to the SaplingNativeBackingStore. I may not need to pass it any further.

Note that I did not copy the path. The "Node" seems to be borrowed data (from the same owner), so it seems okay to also put the RelativePathPiece directly in.

Reviewed By: quark-zju

Differential Revision: D74365297

fbshipit-source-id: 02175b3d7885724eb0dfd999adf8889703313556