pyrepo: add a free-form 'volatile_state' object
Summary:
Various places wants to keep some states on the `repo` object.

Some of them use `ui` configs. Some of them set Python attributes on the Python
`repo` object. However, the native `repo` object does not support setting
attributes on it (not a heap type), and the config of the `repo` is immutable
for now (`Arc<dyn Config>`). To be able to make related code work on native
 Rust repo object, let's add an empty `volatile_state` PyDict to keep track of
 random states.

Reviewed By: muirdm

Differential Revision: D74336718

fbshipit-source-id: ac6b5b9051da3808740ab46da8d602ee75b2e0a1