config: support ui.style being unset
Summary:
In the next change I'm moving `coreconfigitem`s to Rust static configs.
One of the subtle differences is that Rust static config currently respects
`HGPLAIN` filtering but the `coreconfigitem` isn't filtered by `HGPLAIN`.

This can make `ui.config('ui','style')` return `None` and break the template
handling. Use an explicit `''` to avoid the issue.

We might want to revisit how to handle HGPLAIN in static configs. For now,
this is a simple fix.

Reviewed By: zzl0

Differential Revision: D74455332

fbshipit-source-id: 42aec54a424a01eb422015a594046967d2cbbcd6