tweakdefaults: drop fixpager
Summary:
It was useful when `$PAGER` env var is used as `pager.pager`. Later, we stop
using the env var. So there is no need to keep the logic.

If the user uses `--config pager.pager=less`, we still set `LESS=FRX` (at least
from Python):

        pagerenv = {}
        for name, value in rcutil.defaultpagerenv().items():
            if name not in encoding.environ:
                pagerenv[name] = value

This removes a config value mismatch:

  pager.pager
   registered as str (default=internal:streampager) at sapling/configitems.py:358
    mismatched default value:  at sapling/ext/tweakdefaults.py:702

Reviewed By: zzl0

Differential Revision: D74452392

fbshipit-source-id: c33de155ba70b3b23d001dcf7b97c445bb74dc7c