walkdetector: add methods for observing "soft" reads
Summary:
Add "file_touched" and "dir_touched" which correspond to observing a "soft" read. This will refresh the TTL of existing walks, but not create new walks.

The purpose of the walkdetector is to detect slow serial fs access (i.e. reads that hit a remote server to get content). When we see this, we can start prefetching the content in the background. However, now that we have prefetched the content, we will no longer witness the slow reads because they hit the locally cached data. This can cause the tracked walk to time out and be GC'd. This is undesireable since the walk detection would then have to start over from a bunch of intermidiate directories which may not coallesce into a single walk anymore.

Address this with the "touched" methods which can be used to refresh a walk (on a cached read access) without triggering a new walk. If there are no expensive remote fetches then it is pointless extra work to trigger a prefetch (since all the data is already available locally).

Reviewed By: quark-zju

Differential Revision: D74365289

fbshipit-source-id: b3a333bb031446cb2965ea2ad96cda765e084dd7