manifest-tree: allow concurrent bfs_iter calls
Summary:
bfs_iter was not designed to be called from multiple threads at once. It uses a thread pool, but ties up all the threads for a duration the bfs_iter operation (regardless of how much work is done).

Fix by making the workers persistent and moving the iteratoin state into the work item published to the workers. This way, the same set of workers can do work for different/concurrent bfs_iter calls.

This also simplifies the shutdown logic. Now shutdown happens naturally when the last BfsContext is dropped (i.e. the final "result send" side of the channel is dropped).

Reviewed By: quark-zju

Differential Revision: D74365292

fbshipit-source-id: 4759e7071b9db44a7f76a8ade43736e62053bbe5