give backing store a "hint" about parent commit
Summary:
I'm adding some logic to the (Sapling) backing store that needs to know the root manifest id (via parent id) in order to prefetch trees/files. Make the parent id available to the backingstore via a "workingCopyParentHint()" hook that gets called when the mount is initialized and during checkout.

I'm using the word "hint" because is isn't well defined what parent id it is. During checkout, we set to the commit we are going to, since that is probably what we would want to prefetch during the checkout (although not necessarily). During initialization (and "reset") we set it to the working copy parent (as opposed to the checked out commit) because the working copy parent is what we will probably want to use for fetching files/trees.

Reviewed By: quark-zju

Differential Revision: D74365296

fbshipit-source-id: 9874b7fc6bdddd0b42568e734262ac08ca599dac