backingstore: propagate file preload count to walk detector
Summary: As we prefetch file content based on walks reported by the detector, we now feed back stats about the file fetches. In particular, we tell the detector how many files we fetched remotely. This is so the detector can tabulate things to know how good the walk detection and prefetching are working together.

Reviewed By: zzl0

Differential Revision: D74484155

fbshipit-source-id: 7410cae8fde4bca1f70e2d8532f87e52259bc557