backingstore: add prefetch "manager" to kick off prefetches
Summary:
Add a prefetch manager which kicks off file/directory prefetches based on active filesystem walks.

It runs in a separate thread, and waits to be kicked when a walk has changed. It checks intermittently for changed walks so it can notice when a walk has stopped (but there hasn't been any activity to cause it to be kicked).

It allows up to 5 concurrent prefetches. We don't want too many, but we want mroe than 1, so 5 seemed like a good number.

The manager holds on to a reference to the "current commit" so it can notice when EdenFS has checked out a new commit. The prefetches need to be relative to the current commit or you will be fetching irrelevant file/tree content.

Reviewed By: quark-zju

Differential Revision: D74365300

fbshipit-source-id: d8bab752bca45e88efff12df4a96138d5528cd73