manifest-tree: check for iteration cancelation
Summary: Previously we would only check if the iteration had been canceled when atempting to send on the result channel. However, if the queue of work is full of stuff for a canceled operation, we would still process it before breaking out early. Improve by checking for cancelation in more spots, namely at the beginning of the work loop, and before publishing more work.

Reviewed By: quark-zju

Differential Revision: D74365306

fbshipit-source-id: 42a3c9ea41832bb5f9d6608b6b5707007947ace9