backingstore: optimize deeper walk prefetches
Summary:
Add some special logic for optimizing when we get a new walk that has the same root dir but a deeper depth than a walk we are currently prefetching.

Previously we would cancel the existing prefetch and start over with the new, bigger prefetch. This has two issues:

- The existing prefetch is canceled even though it had more useful stuff to fetch
- The new, deeper walk redoes all the fetching from the previous walk. The file should be in local cache, but it is still extra work that will slow things down a bit.

If we notice we have a deeper walk and we have room for a new prefetch, we _dont_ cancel the existing walk, and we start the deeper walk with an offset so it skips prefetching files covered by the previous walk.

Reviewed By: quark-zju

Differential Revision: D74365280

fbshipit-source-id: fc02cdbd2f3445d7ad808130fb9dae00463579dc