walkdetector: start tracking per-walk metrics
Summary:
It will be useful to track the size of walks (i.e. count of file/dir accesses under that walk).

The tabulation is not all hooked up yet - I just added the counters and made things compiles for step one. There was some fallout due to the Walk struct not being Copy or PartialEq anymore.

I went with atomic ints because I plan to eventually split the main Mutex into a RwLock and these counters will be updated during the "read" path.

Reviewed By: quark-zju

Differential Revision: D74365308

fbshipit-source-id: 0ad0a47c8fb4c976f5b9bbe5a3a4b47c0b66931e