walkdetector: add logging for big walks
Summary: Add tracing events for walks that loaded more than 10k files/dirs. We log when they first cross the 10k threshold, and then we also log when they end. I wanted to log when they start so we can get an idea of elapsed time, but also because I'm afraid walks will get "stuck" walking and I want to be able to see that in the logs.

Reviewed By: quark-zju

Differential Revision: D74365291

fbshipit-source-id: 48996ce45cfcad6a6f87ed935e5784bfb79f0817