scmstore: optimize the TreeStore get_content method
Summary:
Add a fast direct-local check for the TreeStore KeyStore::get_content method. This method is used in heavy operations in the manifest-tree crate that look like the below (conceptually):

```
store.prefetch(<lots of keys>);

for item in work {
  store.get_content(item.key)
}
```

Since the key will almost always be in the local caches after the prefetch, it makes sense to do an optimized single-key check.

If we don't find it in local caches, we remove "LOCAL" from the fetch mode, so we skip the local work when we fall back to the heavy weight fetch_batch.

Reviewed By: quark-zju

Differential Revision: D74365283

fbshipit-source-id: 672772c7b722bc9caa3270b1189193d3fee05590