walkdetector: incorporate knowledge of "preloads"
Summary:
Typically when a walk is detected, something will start prefetching content to make the walk faster. In order to better understand the prefetching aspect, the walk detector now supports recording a count of prefetched files, incorporating it into its stats.

In order to understand the efficacy of the walk+prefetch, we need to know how much content was preloaded so the walk detector can differentiate local file reads that were already local before the walk, and local file reads that were just preloaded because the walk was detected.

It is also important to detect cases where we might be over-prefetching (i.e. file_preload_count >> file_read_count).

Reviewed By: zzl0

Differential Revision: D74484157

fbshipit-source-id: 2a03faa1f4c82e00639821b12ca6055aba514bd5