walkdetector: switch to atomic type to track "last_accessed"
Summary:
I want to update this field in the read-only path, so switch to an atomic type.

Instead of using SystemTime, I use a lazy Instant as the "base" epoch value. Didn't benchmark specifically, but I wanted to keep monotonic time and a foolproof API (i.e. static epoch time).

Reviewed By: quark-zju

Differential Revision: D74365275

fbshipit-source-id: a7fd4bc3b376caefb96cff0aec745ead5c714b83