Select a few tests to run in no-network mode
Summary:
## Context

There is a non-trivial cost in running tests in no-network mode.

First of all there is the  maintenance cost: no-network tests are very sensitive to noise from C++ libraries; they also tend to run much slower. These add app to very frequent breakages. Our default response has been to disable no-network mode for tests that fail frequently, which kinda defeats the point.

There is also the CI cost, which is not negligible in the current state of fbcode CI. This translates into our diffs taking longer to pass tests.

## This stack

Let's switch the approach: hand-pick a few tests that we know are high signal, and do our best to keep them healthy.

## This diff

Explicitly enable no-network mode for some tests.

This is intentionally a no-op diff so we can have the conversation about the exact set of tests to tag. A follow up diff will change `dott_test` so this takes effect.

 I selected them following my guts feeling based on the names, but I didn't investigate in depth what they are doing. Please call out any you disagree with.

Reviewed By: lmvasquezg

Differential Revision: D74580776

fbshipit-source-id: 7a220cbe87b83f3374d3c66db2e007c4526ecd8f