Benchmark read_dir()
Summary:
liubov-dmitrieva had a good suggestion to instrument `read_dir()` in the traversal benchmark, just like we instrument `open()` and `read()`. This gives us further lower level details on the performance that calling third-party tools like `grep` and `rg` do not.

This already lead to actionable insights in finding slowness of read() in several configurations of CASC.

Reviewed By: jdelliot

Differential Revision: D74509191

fbshipit-source-id: 664ee12b1007e96d5f75f7b695caf63ef9810bd5