telemetry: add a simple tracing->scuba adapter
Summary:
Add a tracing subscriber that will listen for eden sample events and send to an eden logger.

This will basically allow `tracing::info!(target: "some-type", ...)` to get logged to edenfs_events.

For now it uses a global list of enabled tracing targets that can only be set once. In the future we can improve configurability as needed.

Reviewed By: jdelliot

Differential Revision: D74507878

fbshipit-source-id: 9841d8333a750a92941d8690454a40a408dc6361