Add use_case_id to EdenFsClient
Summary:
# This Diff

Adds a new field/parameter to EdenFsClient creation flow to capture a use-case-id. This use-case-id will be used in future diffs to:
1. retrieve a configuration (oncall, concurrency limits, etc) from configerator
2. track per-use-case usage

# Next Steps
1. Add a use case config struct and store an Arc wrapped instance in EdenFsInstance and EdenFsClient
2. Populate the fields for this use case config with data read from configerator.
3. Use use case config to provide use-case-id where needed and set ohter config options.
4. Leverage all of the above to begin to add use-case (and later session) telemetry

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74579698

fbshipit-source-id: 7ba82b1d252d12add7badf776ca7ad806e35403b