Add use_case_id to EdenFsInstance
Summary:
# This Diff

Adds a new field/parameter to EdenFsInstance creation flow to capture a use-case-id. This use-case-id will be used in future diffs to:
1. retrieve a configuration (oncall, concurrency limits, etc) from configerator
2. track per-use-case usage

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74577262

fbshipit-source-id: 8dff2c128cd686c3ff4ce3902420023e42deb249