Use UseCase to control concurency limits of EdenFS client lib usage
Summary:
# This Diff

Replaces the explicit usage of an optional semaphore with the defined concurency limits found in UseCase.

# Next Steps
1. Populate the fields for this use case config with data read from configerator.
2. Leverage all of the above to begin to add use-case (and later session) telemetry

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74644790

fbshipit-source-id: 63190cefd88896e911c7473f25c2d56b3453ca31