Add use case config struct
Summary:
# This Diff

Add a use case config struct and store an Arc wrapped instance in EdenFsInstance and EdenFsClient and use throughout.

# Next Steps
1. Populate the fields for this use case config with data read from configerator.
2. Use use case config to provide use-case-id where needed and set ohter config options.
3. Leverage all of the above to begin to add use-case (and later session) telemetry

# Context

We would like to be able to know which partner is using our client library and collect metrics. We also want to be able to associate various configuration elements to a use case. To this end we are introducing a use-case-id, use-case configuration, and telemetry collection for a use-case.

Reviewed By: MichaelCuevas

Differential Revision: D74581247

fbshipit-source-id: f0ffe159631c70522489e1c91689f1a1f8d817c0